Shapella configuration changes
