+++ pkg/config.go
@@ -17,22 +17,24 @@ import (
 
 const (
 	GenesisForkVersionMainnet = "0x00000000"
-	GenesisForkVersionKiln    = "0x70000069" // https://github.com/eth-clients/merge-testnets/blob/main/kiln/config.yaml#L10
 	GenesisForkVersionRopsten = "0x80000069"
 	GenesisForkVersionSepolia = "0x90000069"
 	GenesisForkVersionGoerli  = "0x00001020" // https://github.com/eth-clients/merge-testnets/blob/main/goerli-shadow-fork-5/config.yaml#L11
 
-	GenesisValidatorsRootMainnet = "0x4b363db94e286120d76eb905340fdd4e54bfe9f06bf33ff6cf5ad27f511bfe95"
-	GenesisValidatorsRootKiln    = "0x99b09fcd43e5905236c370f184056bec6e6638cfc31a323b304fc4aa789cb4ad"
-	GenesisValidatorsRootRopsten = "0x44f1e56283ca88b35c789f7f449e52339bc1fefe3a45913a43a6d16edcd33cf1"
-	GenesisValidatorsRootSepolia = "0xd8ea171f3c94aea21ebc42a1ed61052acf3f9209c00e4efbaaddac09ed9b8078"
-	GenesisValidatorsRootGoerli  = "0x043db0d9a83813551ee2f33450d23797757d430911a9320530ad8a0eabc43efb"
-
 	BellatrixForkVersionMainnet = "0x02000000"
-	BellatrixForkVersionKiln    = "0x70000071"
 	BellatrixForkVersionRopsten = "0x80000071"
 	BellatrixForkVersionSepolia = "0x90000071"
 	BellatrixForkVersionGoerli  = "0x02001020"
+
+	CapellaForkVersionRopsten = "0x03001020"
+	CapellaForkVersionSepolia = "0x90000072"
+	CapellaForkVersionGoerli  = "0x03001020"
+	CapellaForkVersionMainnet = "0x03000000"
+
+	GenesisValidatorsRootMainnet = "0x4b363db94e286120d76eb905340fdd4e54bfe9f06bf33ff6cf5ad27f511bfe95"
+	GenesisValidatorsRootRopsten = "0x44f1e56283ca88b35c789f7f449e52339bc1fefe3a45913a43a6d16edcd33cf1"
+	GenesisValidatorsRootSepolia = "0xd8ea171f3c94aea21ebc42a1ed61052acf3f9209c00e4efbaaddac09ed9b8078"
+	GenesisValidatorsRootGoerli  = "0x043db0d9a83813551ee2f33450d23797757d430911a9320530ad8a0eabc43efb"
 )
 
 // Config provides all available options for the default BeaconClient and Relay
@@ -54,10 +56,13 @@ type Config struct {
 	RelayHeaderMemoryPurgeInterval time.Duration
 
 	// private fields; populated during validation
-	builders              map[structs.PubKey]*builder
+	builders map[structs.PubKey]*builder
+
 	GenesisForkVersion    string
 	GenesisValidatorsRoot string
-	BellatrixForkVersion  string
+
+	BellatrixForkVersion string
+	CapellaForkVersion   string
 }
 
 func (c *Config) Validate() error {
@@ -75,22 +80,22 @@ func (c *Config) validateNetwork() error {
 		c.GenesisForkVersion = GenesisForkVersionMainnet
 		c.GenesisValidatorsRoot = GenesisValidatorsRootMainnet
 		c.BellatrixForkVersion = BellatrixForkVersionMainnet
-	case "kiln":
-		c.GenesisForkVersion = GenesisForkVersionKiln
-		c.GenesisValidatorsRoot = GenesisValidatorsRootKiln
-		c.BellatrixForkVersion = BellatrixForkVersionKiln
+		c.CapellaForkVersion = CapellaForkVersionMainnet
 	case "ropsten":
 		c.GenesisForkVersion = GenesisForkVersionRopsten
 		c.GenesisValidatorsRoot = GenesisValidatorsRootRopsten
 		c.BellatrixForkVersion = BellatrixForkVersionRopsten
+		c.CapellaForkVersion = CapellaForkVersionRopsten
 	case "sepolia":
 		c.GenesisForkVersion = GenesisForkVersionSepolia
 		c.GenesisValidatorsRoot = GenesisValidatorsRootSepolia
 		c.BellatrixForkVersion = BellatrixForkVersionSepolia
+		c.CapellaForkVersion = CapellaForkVersionSepolia
 	case "goerli":
 		c.GenesisForkVersion = GenesisForkVersionGoerli
 		c.GenesisValidatorsRoot = GenesisValidatorsRootGoerli
 		c.BellatrixForkVersion = BellatrixForkVersionGoerli
+		c.CapellaForkVersion = CapellaForkVersionGoerli
 	default:
 		network, err := c.readNetworkFromConfig(c.Network)
 		if err != nil {
@@ -99,6 +104,7 @@ func (c *Config) validateNetwork() error {
 		c.GenesisForkVersion = network.GenesisForkVersion
 		c.GenesisValidatorsRoot = network.GenesisValidatorsRoot
 		c.BellatrixForkVersion = network.BellatrixForkVersion
+		c.CapellaForkVersion = network.CapellaForkVersion
 	}
 	return nil
 }
@@ -107,6 +113,7 @@ type Network struct {
 	GenesisForkVersion    string `json:"GenesisForkVersion"`
 	GenesisValidatorsRoot string `json:"GenesisValidatorsRoot"`
 	BellatrixForkVersion  string `json:"BellatrixForkVersion"`
+	CapellaForkVersion    string `json:"CapellaForkVersion"`
 }
 
 func (c *Config) readNetworkFromConfig(network string) (Network, error) {
