--- pkg/utils.go
@@ -1,67 +1,0 @@
-package relay
-
-import (
-	"errors"
-
-	"github.com/ethereum/go-ethereum/common"
-	"github.com/ethereum/go-ethereum/common/hexutil"
-	"github.com/flashbots/go-boost-utils/types"
-	"golang.org/x/exp/constraints"
-)
-
-const Version = "0.3.6"
-
-type UserAgent string
-
-// ComputeDomain computes the signing domain
-func ComputeDomain(domainType types.DomainType, forkVersionHex string, genesisValidatorsRootHex string) (domain types.Domain, err error) {
-	genesisValidatorsRoot := types.Root(common.HexToHash(genesisValidatorsRootHex))
-	forkVersionBytes, err := hexutil.Decode(forkVersionHex)
-	if err != nil || len(forkVersionBytes) > 4 {
-		err = errors.New("invalid fork version passed")
-		return domain, err
-	}
-	var forkVersion [4]byte
-	copy(forkVersion[:], forkVersionBytes[:4])
-	return types.ComputeDomain(domainType, forkVersion, genesisValidatorsRoot), nil
-}
-
-type GetValidatorRelayResponse []struct {
-	Slot  uint64 `json:"slot,string"`
-	Entry struct {
-		Message struct {
-			FeeRecipient string `json:"fee_recipient"`
-			GasLimit     uint64 `json:"gas_limit,string"`
-			Timestamp    uint64 `json:"timestamp,string"`
-			PubKey       string `json:"pubkey"`
-		} `json:"message"`
-		Signature string `json:"signature"`
-	} `json:"entry"`
-}
-
-func Max[T constraints.Ordered](args ...T) T {
-	if len(args) == 0 {
-		return *new(T) // zero value of T
-	}
-
-	if isNan(args[0]) {
-		return args[0]
-	}
-
-	max := args[0]
-	for _, arg := range args[1:] {
-
-		if isNan(arg) {
-			return arg
-		}
-
-		if arg > max {
-			max = arg
-		}
-	}
-	return max
-}
-
-func isNan[T constraints.Ordered](arg T) bool {
-	return arg != arg
-}
