+++ pkg/validators/validators_test.go
@@ -13,6 +13,7 @@ import (
 	"github.com/blocknative/dreamboat/pkg/validators"
 	mock_val "github.com/blocknative/dreamboat/pkg/validators/mocks"
 	"github.com/blocknative/dreamboat/pkg/verify"
+	"github.com/blocknative/dreamboat/test/common"
 	"github.com/flashbots/go-boost-utils/bls"
 	"github.com/flashbots/go-boost-utils/types"
 	"github.com/golang/mock/gomock"
@@ -32,7 +33,7 @@ func TestGetValidators(t *testing.T) {
 	ver := verify.NewVerificationManager(l, 20)
 	ver.RunVerify(300)
 
-	relaySigningDomain, err := pkg.ComputeDomain(
+	relaySigningDomain, err := common.ComputeDomain(
 		types.DomainTypeAppBuilder,
 		pkg.GenesisForkVersionRopsten,
 		types.Root{}.String())
@@ -67,7 +68,7 @@ func TestRegisterValidator(t *testing.T) {
 	ds := dbadger.NewDatastore(store, time.Minute)
 	bs := mock_val.NewMockState(ctrl)
 
-	relaySigningDomain, err := pkg.ComputeDomain(
+	relaySigningDomain, err := common.ComputeDomain(
 		types.DomainTypeAppBuilder,
 		pkg.GenesisForkVersionRopsten,
 		types.Root{}.String())
@@ -121,7 +122,7 @@ func TestBrokenSignatureRegisterValidator(t *testing.T) {
 	ds := dbadger.NewDatastore(store, time.Minute)
 	bs := mock_val.NewMockState(ctrl)
 
-	relaySigningDomain, err := pkg.ComputeDomain(
+	relaySigningDomain, err := common.ComputeDomain(
 		types.DomainTypeAppBuilder,
 		pkg.GenesisForkVersionRopsten,
 		types.Root{}.String())
@@ -188,7 +189,7 @@ func TestNotKnownRegisterValidator(t *testing.T) {
 	ds := dbadger.NewDatastore(store, time.Minute)
 	bs := mock_val.NewMockState(ctrl)
 
-	relaySigningDomain, err := pkg.ComputeDomain(
+	relaySigningDomain, err := common.ComputeDomain(
 		types.DomainTypeAppBuilder,
 		pkg.GenesisForkVersionRopsten,
 		types.Root{}.String())
@@ -236,7 +237,7 @@ func BenchmarkRegisterValidator(b *testing.B) {
 	ds := dbadger.NewDatastore(store, 5*time.Minute)
 	bs := mock_val.NewMockState(ctrl)
 
-	relaySigningDomain, _ := pkg.ComputeDomain(
+	relaySigningDomain, _ := common.ComputeDomain(
 		types.DomainTypeAppBuilder,
 		pkg.GenesisForkVersionRopsten,
 		types.Root{}.String())
@@ -284,7 +285,7 @@ func BenchmarkRegisterValidatorParallel(b *testing.B) {
 	ds := dbadger.NewDatastore(store, 5*time.Minute)
 	bs := mock_val.NewMockState(ctrl)
 
-	relaySigningDomain, _ := pkg.ComputeDomain(
+	relaySigningDomain, _ := common.ComputeDomain(
 		types.DomainTypeAppBuilder,
 		pkg.GenesisForkVersionRopsten,
 		types.Root{}.String())
