+++ scripts/getpayload/main.go
@@ -11,6 +11,7 @@ import (
 	"github.com/blocknative/dreamboat/blstools"
 	relay "github.com/blocknative/dreamboat/pkg"
 	"github.com/blocknative/dreamboat/pkg/api"
+	"github.com/blocknative/dreamboat/test/common"
 	"github.com/ethereum/go-ethereum/common/hexutil"
 	"github.com/flashbots/go-boost-utils/types"
 	"github.com/pkg/errors"
@@ -33,7 +34,7 @@ func main() {
 }
 
 func submitRequest() error {
-	builderDomain, err := relay.ComputeDomain(
+	builderDomain, err := common.ComputeDomain(
 		types.DomainTypeAppBuilder,
 		relay.GenesisForkVersionRopsten,
 		types.Root{}.String())
